refactor(cli): drop yargs-util
