+++ lib/yargs/middlewares/trace-locations-from-argv.js
@@ -1,14 +1,17 @@
 'use strict'
 
-const {
-  getLocationDataFromArgv,
-  getVendorDataFromArgv
-} = require('../yargs-util')
-
 const middleware = () => {
   return argv => {
-    const locationData = getLocationDataFromArgv(argv)
-    const vendorData = getVendorDataFromArgv(argv)
+    const locationData = {
+      description: argv.description,
+      address: argv.address,
+      startTimestamp: argv.startTimestamp,
+      endTimestamp: argv.endTimestamp
+    }
+    const vendorData = {
+      type: argv.type,
+      defaultCheckInLengthInMinutes: argv.defaultCheckInLengthInMinutes
+    }
     argv.__dataFromArgv = [
       { locationData, vendorData, filepath: argv.filepath }
     ]
